refactor: ページウィジェットを/pagesにまとめた
